Add files via upload
The main objective of this project is to investigate the feasibility of utilizing
accelerometer data from the MXCHIP Devkit for predicting hand movement and detecting
writing. The collected and labeled accelerometer data were used to train a decision tree
classifier for the analysis. The purpose of this research is to identify potential virtual network
techniques that can enable the implementation of Internet of Things (IoT) devices in diverse
applications.
Our dataset was
limited to only one participant, which may not be representative of the wider population.
Furthermore, we only focused on writing as a specific type of hand movement, and future
research could investigate the potential for detecting other types of movements. In the future,
we plan to extend our work to predict not only whether the user is writing, but also what they
are writing. This could involve using more advanced machine learning techniques such as
recurrent neural networks or convolutional neural networks.
Overall, our research demonstrates the potential for using accelerometer data from
wearable devices to predict hand movement and detect writing. While there are limitations to
our study, our findings contribute to the growing field of wearable technology and have
practical applications in areas such as user experience design and assistive technology.
Further research is needed to explore the potential for detecting other types of hand movement
and to validate our findings on a larger and more diverse sample